requirements.txt
django >= 1.11
djangorestframework
django-crispy-forms
django-rest-auth
git+https://github.com/carltongibson/django-filter.git
django-expvar-psutil